Change pdf page count method output
The getPageCount method now returns a float array of size 2. The first
entry is the maximum width of the pdf amoungst any of its pages. The
second entry is the total height of all the pages together. This will
allow us to dymanically create a scrollable window for any size pdf