Create DetailedAckDocView
After a discussion with Sam, and discovering that PDDocument is not
thread safe, we decided to optimize the amount of times that the
PDDocument objects load in a file. Additionally, these methods are only
needed in certain situations and because of that we can seperate them from
the main AckDocView. Hence, the DetailedAckDocView was created