Display ack doc in embed for non edge
Internet explorer does not scroll well when the ack doc is loaded with
an iframe.  We cannot fully move back to embed because that doesn't work
in edge.  The solution is to use an iframe tag only if the user is in
edge.

This commit imports a new third party library, bowser, to assist
in detecting the browser of the current user.