Simplify layout of document acknowledge page
Users were confused by the document acknowledgment page because the "I
Agree" button that is clicked to acknowledge the document is not visible
when the page loads and must be scrolled to see.  The users reported not
being sure what to do on the page because the objective was not visible.

I have addressed this issue by reducing the height of the instructions
and acknowledgement sections of the page, eliminating the necessity for
scrolling the top level page for most reasonable window sizes.

Refs #11695