Update acknowledge page design
Alters the layout of the acknowledge page to remove the need for
multiple scrollbars.  In the new design, instructions and document are
expanded to fit and scroll with the page as a whole.  The acknowledge
button is fixed to the bottom of the viewport, ensuring that the user
can always see it, no matter their screen size.

This new design is enabled by an css hack involving blank iframes
that allows us to render elements over a pdf container in IE.