Save Work related to report refactoring
After the meeting with personnel, I began refactoring the reports to
what we agreed on. The report that operates by ack doc id does not throw
any errors but the data that is returning isn't 100% correct at the
moment. No exceptions are thrown which makes the situation even more
interesting