Isolate login angular app.
The login page does not need most of the functionality provided by the
ESS angular app.  There are also some features of the main angular app
(such as keepalive pings) that are enabled by default and cause issues
in the login page since the user is not authenticated.

This commit moves the login page angular modules into a distinct angular
app, which imports a minimal amount of the main ess angular app.