Track session timeout on the backend.
The current session timeout mechanism works pretty much entirely on the
front end.  It consists of an angular component that counts down the
user's inactivity and logs them out when they are inactive past the
timeout duration.  The actual session timeout is disabled on the server
and will persist the session until the user explicitly logs off.  This
allows the user to bypass the timeout if they close their browser window
before the timeout ends.  When they go back to ESS, their session will
not have expired and the front end timeout counter will restart at 0.
Also, the design of the timeout mechanism is dubious and could be
improved greatly.

This commit overhauls the way that sessions are timed out.  Instead of
tracking inactivity on the front end, user inactivity is now tracked on
the back end by storing the timestamp of the last active ping as a
session attribute.

The front end component will watch for user activity and send a ping to
the server every 30 sec.  If the user was active in the last 30 sec, the
active parameter will be set to true for the request.  If the parameter
is true, the last active ping session variable will be updated.  The
ping api request will return the amount of inactivity allowed before the
user times out, allowing the front end to display a warning modal when
this gets too low.

The last active ping session attribute comes into play via security
filters on page and api requests.  If enough time has elapsed since the
last active ping timestamp, the page filter will log the user out and
redirect them to the login page.  In this case the api filter will
return 401.