Fix Doubling search results console errors
The employee search directive would double the employees when the select
another employee was clicked. I originally fixed this by calling a
digest cycle which caused console errors.

The reason why the digest cycle worked, was because it was resetting the
variable which is what I did right before it gets modified again.