Grant ack managers permission to view emp info
Employee's with the ACK_MANAGER role, but not the TIME_MANAGER role
were getting authorization errors when using the employee ack report
page.  This was because the employee search directive was attempting to
access employee info.

This commit grants permission to access basic info for all employees to
users with the ACK_MANAGER role.

Fixes #11846