Fix bug in employee search directive.
There is a bug in the employee search directive where the previous
results are cleared when the user scrolls down and triggers a new page
of searches via the infinite scroll api.  This bug was introduced as a
fix to another bug where duplicate search results would be shown when
the "Select another employee" button is pressed.

The original problem was due to the fact that a new search was
performed when the reselect button was pressed and the empId parameter
was not cleared, so the search would return results for the same
employee.

This commit resolves both of these issues by not performing a search
when the button is pressed and clearing the empId variable/request
param.  The search is still performed if the user accessed the page
using the empId param because in this case, no search has been performed
yet.