Add ability to submit acks from ack report page.
This commit modifies the employee acknowledgment report page by adding
the ability for personnel ack. managers to submit acknowledgments on
behalf of an employee.  Any pending acknowledgments on the employee
will open an acknowledgment prompt if clicked.  If this prompt is
accepted, an acknowledgment will be posted for selected doc/employee.

Refs #11714