Return user to last page after timeout.
This commit alters the front end timeout logic so that the user is
returned to the page they were on when they timed out after logging in
again.  This is achieved by logging them out via an ajax request, and
then reloading the current page, triggering shiro's "save page and
redirect to login" feature for unauthenticated users.