Allow new time recs to overwrite overlapping recs
There have been issues where the time record manager cannot insert a
new time record due to an existing time record on the same date range.
There is already logic that allows for overwriting an existing time
record with the same begin date, but this logic fails if the begin date
is not the same.

This commit modifies the query that checks for an existing time record,
allowing it to return time records with any overlapping dates instead of
just a matching being date.

Fixes #11904