Fix entry patching bug in time record manager.
There was a bug in the time record manager that caused records whose
entries needed to be patched not to be saved.  For some reason, the
condition for saving a record if it was not split or deleted was that
there was a record data modification AND an entry modification.  This is
a redundant check since the record data modification method also
modifies entries.  Because the entries are modified in the
patchRecordData method, the subsequent call of the patchEntries method
will always return false, because the entries are already patched.

This bug is fixed by removing the redundant calls to the patchEntries
method.

Fixes #11925