Fix time record last update query
The query that gets the update date of the last updated time record was
ignoring time records without time entries due to a join on the entry
table.  I have fixed this by switching it to a left join, allowing time
records without time entries to appear in the results.