Factor in AP records in time record manager.
The code in the time record manager was running under the assumption
that all personnel approved records had a corresponding attendance
record.  This is not the case and has caused personnel approved time
records to be overwritten because the TRM thought there was no record
during their range.

This commit enhances the time record manager by making it factor in the
personnel approved records during time record range calculation.
Personnel approved records are now removed from the set of ranges for
which time records are expected.

Refs #11948