Wrap TA report in ess api
This modifies the ESS front end so that instead of providing a link
directly to the sfms biweekly time and attendance report on the accrual
pages, the link points to an ess api report url that wraps the sfms
report.

This is done for two reasons:

- Hour values need to be passed to the sfms report via url params.  If
  employees notice this, they could forge fake reports.
- This allows the logic of generating the report param values to be done
  on the ESS backend.

The backend rewrite of the report url generator also addresses an issue
where the reports did not show an employees work hours if the report was
generated between the time the emplyees attend record was posted and the
time the corresponding acc usage record was posted.

Fixes #11944