Handle change of non-work RA/SA entries to TE.
When an employee's pay type changes from regular or special annual to
temporary, the change may be retroactive over already entered time, or
may cover future days where accrual, misc or holiday hours are entered.
The existing time record manager behavior was to preserve these entries,
but change their pay type.  This is invalid, because temporary employees
are not allowed to be paid for non-work hours.

This commit addresses this issue by modifying the time record manager to
deactivate any temporary pay entries containing non-work hours.  This
preserves the data from these entries in case there is a mistake, but
removes them from consideration and allows the user to overwrite them.

Fixes #11966