Fix date errors in accrual compute service.
The accrual compute service was not using the fromDate parameter
correctly in two instances.  Both misuses were instances where the
fromDate parameter was treated as an inclusive end date rather than a
begin date.  The from date is a begin date.  It is either the day after
the employee's last annual accrual record the first day of the current
continuous block of service if they do not have an annual accrual
record.

In the first instance, the method that returns gap accruals was
returning empty if the from date was greater than or equal to the last
period end date.  This is incorrect in the case where the employee is
active for only the last day of the period because the from date = end
date in that case.  I modified this to return empty if the from date is
after the begin date.

In the second instance, the from date is set as the end date of the
initialized accrual state object.  I modified this to set the end date
as the day before the from date.