Implement basic validation for travel application
This implements validation on the purpose, outbound, return, and
allowances page of the travel application.

I say this is basic because some of the most important inputs - address
and date - only validate that the input is not empty. Will need to create
better validators for these inputs in a future commit.

refs #11995