Incorporate leave exception in accrual calculation
There is a special rule for accrual computation that was not present in
the accrual calculation logic.  This rule is that employees are not
allowed to accrue time if their personnel status is sick leave with half
pay or leave without pay on the last day of a pay period.

This commit adds this logic into the accrual compute service.

Fixes #12042