Add more external api failure modals to travel app
Fixes #12096

During the travel application, ess communicates with a few outside
services like Google and GSA to have all of the information for the
estimate. If we cant connect to these services, then the travel app
should fail gracefully. It will now show a modal and allow the user ot
go back to ess or log out of travel