Add Meal Rates Scrapper back into code
The meal rates scraper and the subsqunet scraping that would occur was
removed from the code base a while back. I have updated the previous
code and it will run on a cron job. The new cron job has a new app
property, cache.cron.meal.rate = 0 0 0 * * *

This new system also has an admin controller to trigger the scrape
manually if need be