Fix bug in time entry deactivation.
There has been an issue with deactivating time entries when the date
range of a time record is reduced.  This is caused by two things:

1. The time record query ignores time entries that do not fall within
the time record's date range.  Time records are always saved before
their entries.  When entries are saved, they are compared against
existing entries to prevent saving of empty or unchanged entries.
When the time record is saved with a reduced date range, the time record
that is retrieved for entry comparison does not contain the entries to
be deactivated.
2. When saving time entries, entries not present in the previous time
record will be saved only if they are non-empty and inactive.

The requirement for retrieved time entries to fall within the time
record's date range has been removed, because this will mask entries
that need to be deactivated.  The active requirement for saved new
entries is also removed.

Fixes #12144