Update ESS dist assign functionality
The controller and service for district assignment was not capable of
handling all of the inputs that are accepted in SAGE for district
assignment. I added some new controller methods, and updated how the
district assignment service creates the url to call so that it always
uses the complete provided input.