Create Dod Scraper v1
For Alaska, Hawaii and other countries we need to get the per diem
information from a source other than GSA. That source happens to be the
DOD. Their website structure must be scraped and it must be manipulated
by a user agent. Using Jsoup we are now able to get the rates of a town
in Alaska, Hawaii etc.
There is still work to be done. We need the ability to get rates at a
previous date of publication, which can be accomplished with some
additional form manipulation with Jsoup