Fix issues with time record save error handling.
There were some issues with error handling in the method that saves time
records.  The save method is Spring Transactional, which means that all
database operations in the method are wrapped in a transaction, which
is not committed if the method produces an exception.  If a time record
is saved and cached during the method, but then an exception occurs, the
database is reverted, but not the cache.  The following changes have
been made to address this:

- Catch exceptions in the save method and clear the time record cache
  when one is caught.
- The same exception handling is added to the time record manager, which
  performs multiple time record saves under a single transaction.
- Trigger notification emails via events rather than in the method to
  prevent email errors from disrupting saves.
- Move time record validation to the save method rather than the REST
  controller.

Fixes #12166