Fix bug with time record truncation.
There was a bug that occurred when time records were shortened due to
employee termination.  Future entries of these time records for days
after the termination were not removed.  This occurred because these
entries were removed from the record in the split/truncate method, and
therefore were not modified in the database.

This commit modifies the split/truncate logic so that entries are only
removed if there is a new split record to move them into.  This allows
truncated entries to be deactivated later in the process.

Fixes #12365