Rework handling of appoint transaction records.
A refactoring of the method used to identify and apply appointment
transaction records in TransactionHistory in response the previous
method's handling of an edge case.

The previous method was not handling a case where the employee's APP
record was retroactively given a later date while the other transactions
associated with the appointment remained at the earlier date.

The primary goal of this refactor was to eliminate this scenario by
removing any snapshots that were created before the first APP record.
Even though the snapshots are removed, the values added from the
transactions prior to the APP are still included in the APP's snapshot.

I also attempted to make the logic of the method clearer and less
redundant.

Fixes #12374