Prevent time record merges.
This is a modification of the time record manager logic to prevent
instances where the TRM attempts to merge two time records.  There is
currently no clean way to do this because one of the two records will
need to be deactivated with the other expanded, and this runs into
unique key problems during saving (keys exist for empid/begin date and
empid/end date).

I also added some logic to address scenarios where approved records are
retroactively given a different supervisor.  Now, when this scenario is
detected, the records in question will be reverted to "Submitted" state,
allowing the new supervisor to approve them.

Fixes #12396