Fix supervisor assignment issues in front-end.
There was an issue with the way the front end was determining the
supervisor of an employee based on a loaded supervisor employee group.
The problem was related to the fact that all of the loaded supervisor
data is point in time, while the method to determine an employee's
supervisor uses a one to one mapping.  It is entirely possible for an
employee to have multiple assigned supervisors for the covered time
period.

To address this, I have removed the dependency on this method as much as
possible:

- Created SecondarySupEmpGroupView class that contains the id of the
  supervisor of the emp group's supervisor.  This is used in the
  ExtendedSupEmpGroup for the emp groups of the primary supervisor's
  employees.
- Use the new secondary emp group field in the front end instead of the
  map.
- Prevent employee overrides from populating the supervisor id map.

Fixes #12437