Fix date comparison issue in accrual service.
There was an issue with date comparison in the accrual compute service.
The issue would occur when submitted time records were being examined to
see if they matched up with a pay period for which accruals need to be
computed.  The code iterates through a sorted list of time records
looking for date range overlap until the next time record begin date is
not after the period end date.  This works fine until the time record
begin date equals the period end date, which be the case if a time
record only covers the last day of the period.  This causes the time
record to be ignored for accrual computation.

This commit fixes the above mentioned bug by changing the comparison so
that it requires the gap period end date to be 'not before' the record
start date, essentially gap end date >= record start date.

Fixes #12611