Add dynamic query system for personnel tasks.
Add another method to the query dao that allows for more freeform
queries using a query object.  The new query will match based on one or
more parameters set in the query object.  Unset parameters are ignored
in the query.  This one method allows for queries such as "get list of
all incomplete tasks for all employees" or "get all completed tasks of a
specific type for a single employee".