Use best match path to set active menu item.
Modifies the left nav code for highlighting the active page.
Previously, only the menu item with an exact path match was set as
active.  Now, the algorithm strips down the path until a match is found,
effectively highlighting the "parent" page if no menu item exists with
the exact same path as current page.