Use task specific APIs to complete tasks.
I initially made the task api to have a general method for saving
completed tasks.  This doesn't work since not all tasks can be completed
in the same way.  This removes that method, and replaces existing usages
(i.e. acknowledgment) with new api methods specific to the task type.
Now, instead of saving a full task to indicate completion, a specialized
DAO method is used which sets completion fields for a task with a given
emp id and task id.