Prevent insertion of inactive time entries.
The insertion of inactive time entries has been causing problems,
because there is a trigger in the time entry table that sets newly saved
rows as active.  Saving inactive entries has been occurring when holiday
entries are deactivated via the TRM and then saved.  Holiday entries
will not be saved to the database until the user submits the time
record, but they will be present on cached time records.  These get by
the filter method that prevents saving of empty time entries (because
they are preset with holiday hours) and are activated by the trigger
when saved.

This commit adds a condition to the time entry dao to prevent insertion
of inactive time entries.  Updates are still allowed, because existing
time entries will need to be deactivated.  Inserts of inactive entries
are unnecessary.

Fixes #12734