Remove ability to review own timesheet in frontend
Remove the ability to for a supervisor to remove their own timesheet on
the time record management page in the event of a cyclical supervisor
relationship.  If a supervisor is supervised by their own employee, then
they will still see their own records on the manage page, but will be
unable to select the records for approval/disapproval.

Fixes #12737