Remove app review permission from regular users
Travel App and Review permissions were part of the CorePermissions but
that was not a good place for them. One side effect is that all users
had permission to read/write to their AppReviews which they are not
allowed to.

This fix creates new travel permission classes and a TravelPermissionBuilder
for constructing them. The builder uses shiro multiple parts to
efficiently store multiple permissions in a single WildcardPermission instance.