Use directive in record history page.
The user record history page had a unique controller and template that
performed virtually the same functions as a record history directive
used on the supervisor record history and employee search pages.  The
user controller had a bug that mishandled the case where a user
submitted a paper timesheet that overrode an electronic timesheet.
I took this opportunity to remove the redundant controller on the user
record history page and replace it with the directive.

Fixes #12712