Training Video CSV Processing
This is code to process the CSV files that contain employee training data.
It requires that the CSV files are in fact in CSV format and are in the
proper directory. Additionally, it requires that the 2 videos (to date)
have database records in the table ess.pec_video. This process can be
kicked off with an api call