Improve supervisor cache updates
The process of updating the supervisor emp group cache in response to a
transaction post was flawed.  The method would get affected supervisors
using the current transaction history.  This can leave out supervisors
in scenarios where the supervisor id was changed on a transaction, or a
supervisor transaction was inactivated.

The method is modified so that it gets affected supervisors using both
the current transactions, and the current cached supervisor emp groups.
This ensures that all affected supervisor emp groups will be evicted
for any transaction scenario.

Refs #12813