Add completion status to personnel task search
Adds a completion status facet to the personnel task search.  This was
implemented slightly different than was planned, without the ability to
search for tasks completed during a time range.  This was because the
results on the page encompass an employee and multiple tasks as opposed
to individual tasks, and it wouldn't be intuitive to display the results
of that query in this manner.

A new query parameter was added on the backend, "totalCompletion", which
allows filtering by the overall completion status of all tasks in an
employee/task search result.