Add an supply item ignore list for the reconciliation process
Adds the supply.reconciliation.ignore.list property which contains a
comma separated list of commodity codes that do not require
reconciliation.

Created a ReconciliationDtoView object which provides the UI all the
data it needs for the reconciliation page. This simplifies the front end
logic but was also necessary to consolidate the reconciliation item list
logic into a single method.

fixes #12276