Remove exceptions due to invalid slack.webhook.url
Replaced the IllegalStateException with a warning message when sending
a slack message without a properly configured slack.webhook.url.

There is no need to throw an exception here as our application should
continue even if it fails to send a message to slack. A warning message
should be enough to notify us of an incorrect url.

This parameter is often purposefully left blank in dev environments.