Fix param serialization for todo report csv link
Modify todo report code to allow for report params to be used in the
link to generate a csv file.  The parameter string was being generated
but was incorrectly referenced in the report template.

Also, move the param string variable into the "state" object for
consistency, and use angular's $httpParamSerializer to generate the
string instead of the jquery param method.