Remove locking of scroll bar when a modal is displayed
Locking the scroll bar had the unintended effect of returning the user
to the top of the page they were on. This can cause users to lose
context when clicking on a modal.

The only side effect of removing the locked scroll bar is that users can
scroll on the background page while a modal is open. I don't see an
issue with this because the entire background stays dark and the modal stays
centered.