Add scroll pagination to rch picker directive
The previous strategy of loading all responsibility heads and filtering
them in the directive was not performing well in IE for the default
unfiltered view that renders all rchs at once.  To mitigate this, I have
reverted to request based filtering that uses pagination.  Only the
first 20 results are loaded, and subsequent results are loaded when the
user scrolls near the bottom of the current results.

To accomplish this, an offset attribute was added to the "on scroll to
bottom" directive, which allows for the trigger to be applied before the
user reaches the bottom.  This happens at a specific number of pixels
above the bottom, specified by the attribute.