Use pdf.js to render acknowledgment docs
Use the pdf.js library to display pdfs in the ack doc view page.  This
is done to reduce bugs from browser differences when natively rendering
pdfs in embed or iframe tags.  This was spurred by a bug in chrome that
made pdf rendering very slow or impossible when the embed tag was
extended to be greater or equal to the document in height.

Many browser specific hacks for pdf scrolling detection are removed in
addition to the "angular-bowser" library which enabled these hacks to be
employed on their target browsers.

A loading indicator is added to the acknowledge button on the page to
account for the time it takes to fully render the pdf with pdf.js.