Add defaults to moodle generation api call
Some default params have been added to the moodle api call to make it
easier for the admin, without losing functionality.

The api call will default to 1970 to tomorrow with a default
organization of the "senate".

Different organizations and time ranges can still be specified as
normal, but the defaults are our primary use case