Fix inactive docs not appearing in completed area
Once an ack doc becomes inactive it was no longer appearing in the
completed section. This was not in accordance with the specifications of
the project.

Now a new user will only see the active ones in the unacknowleged
section where as employees who have been here longer will see previous
inactive docs if they have ackownleged them