Change cron for task assignment
The task assigner will now run at 3am instead of 1am. This was changed
so that hopefully a Out of memory error doesn't occur when ESS needs to
refresh other parts of its data at the same time.

It is possible this change will not change anything, but it won't hurt
either