Add ability to manually update personnel tasks
The ability to manually update a personnel assigned task for a given
employee through the reporting interface has been added.

This is ideal because it takes advantage of the search mechanisms
already present. This will allow personnel to make manual updates
without having to contact us.

The only downside is that after a manual confirmation, the report
refreshes and to get back to the same spot they must reenter the search
criteria.

With that said, this is a feature they should not have to use very often
if at all