Save work for updating bad email & empids
I am committing my work for updating bad emails and bad employee id data
in various everfi user records.

The part that would actually contact everfis prod database is commented
out while I continue to test it. Some of this is needed for another part
of the maintaince of everfi so it must be committed.