Change Assert for insertUpdateTest
The previous assert appeared to be failing because of some change from
J8 to J15. Specifically in comparing local date time objects.

The test is meant to see if the dao updated the database, so the test now checks
to see if the completed part of the update is set to true, which would
indicate a successful update since that value would be null before.