Use consistent precision for requisition modified date time
This fixes a bug with Supply's pessimistic locking caused by a
requisition being saved to the database with a modified date time
in nanosecond precision yet the API response only containing  microsecond
precision. If someone tried to update this requisition the pessimistic
lock check would fail because the modified date time values did not
match.

refs #13975