Everfi User Fixes
The report email did not respect the new line, and the data will now be
a little more clear.

Additionally, new users will be skipped if there is no email present.
They will still appear in the report email so we know who was attempted
to be pulled into Everfi