Fixes missing columns in paycheck history
Recent changes to the deductions applied to paychecks caused some users
to have multiple deductions with the same description in their paycheck
history view. This view was programming assuming descriptions were
unique, and therefore it was failing.

This fix adjusts the code to check deduction uniqueness based on their
code field instead of their description. Additionally, lots of logic
was moved from the front end into the backend with the PaychecksSummary
class and implementation of the Comparable interface in the Paycheck
and Deduction classes.

See issue #13884 for more background info on the changes to deductions.