Fix null check 4 email when adding users to Everfi
A check to ensure employees had a valid email before being added to
Everfi was in reverse. This was causing a null pointer exception and
that would cause a failure and prevent the process from continuing.
Additionally, a few other logging points in the Everfi User Service have
been updated to be more clear for future use