Fix bug in mileage reimbursement calculation
All methods of travel should be used in the calculation of the outbound
mileage distance. Previoulsy we only checked outbound legs with methods
of travel that qualify for reimbursement.

This fixes an issue where a user taking a train to their destination but
driving home was not receiving any mileage reimbursement.