Fixes time exceptions for Parmely
Parmely has been employed since before SFMS, therefore her continuous
service date predates her SFMS time records.

We were encountering an edge case where her Annual Accrual endDate
was null which cause a calculation to fall back to the continuous
service date. For most employees this works, but it fails for her
since there are no records from when she was first hired.

Now, in instances of a null Annual Accrual end date, we use either the
continuous service date or the last day of the previous year, whichever
is greater. This is the only timeframe that is necessary for calculating
the AccrualState.

refs #15280