Better First Invite tracking
Changes have been made to ESS PEC to allow for the tracking of first
notifications and to ensure that we dont send out invite emails more
than once. Notifs will not be marked complete in test mode.

Additionally, the exclusion serivce was wired in to the auto assignment
process as some employees have different training requirements and
should not be tracked like most employees.

Finally, a bug was fixed relating to querying the manual override status
of any individual task assignement before we insert it into the database