Change query time to everfi
Our Everfi records were not always been generated or updated properly.
This is due to a cap on the number of requests we can send in a time
period.

By changing the time period that we receive results for the problem is
eliminated. This also works well, sine we only care about our most
recent trainings.

Additionally, the Everfi cron was not bound by sync being enabled. This
will prevent any other environment from accidentally taking up our
limited requests to Everfi. This cron was always supposed to be limited
by this app property and that has been corrected