Improve the speed of employee cache warming
This commit modifies the employee cache to query all employees at once
instead of querying them each individually. This reduces the time needed
to initialize all employees from ~95s to ~14s.

This improvement is needed in for the Travel v2 code where all employees
are loaded from the cache to determine permissions on login.
Without this improvement it takes ~2min to login after booting up the
app, which when developing and testing is unreasonable.