Add back resource filtering to the pom.xml
This was causing issues with a spreadsheet file in our resources
directory, and we couldn't see where it was doing anything, so we
removed it.

Turns out it is used in versioning our js files, and without it browsers
do not clear their caches on version updates like intended.

Without this config, our js files are named:
ess-vendor.min.js?v=@project.version@

But they should be named something like:
ess-vendor.min.js?v=1.9.2

The spreadsheet file which was causing issues has been moved into our
wiki documentation which is a more appropriate location for it.